refactor(engine): move HTTP error mapping out of contracts/
Addresses James's review: contracts/ should only hold the canonical
seam, not backend-specific logic. Moves HTTP status dispatch, URL
sanitization, shell quoting, and provider-detail enrichment from
contracts/query_error.rs to backends/http/error.rs alongside
ProviderQueryError. The mapping is now
ProviderQueryError::to_structured() → StructuredQueryError.

contracts/query_error.rs is now just the struct definition, Display
impl, getters, and a generic new() constructor.

Also fixes 400 → InvalidArgument (was FailedPrecondition).
